Consolidate geo package
