+++ app/app.go
@@ -12,8 +12,8 @@ import (
 	"github.com/gdamore/tcell/v2"
 	"github.com/gdamore/tcell/v2/views"
 	"github.com/jensgreen/dux/dux"
+	"github.com/jensgreen/dux/geo/z2"
 	"github.com/jensgreen/dux/nav"
-	"github.com/jensgreen/dux/z2"
 )
 
 type App struct {
