+++ geo/z2/interval_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/jensgreen/dux/r1"
+	"github.com/jensgreen/dux/geo/r1"
 )
 
 func Test_Interval_Contains(t *testing.T) {
