+++ treemap/tiling/tiling.go
@@ -2,8 +2,8 @@ package tiling
 
 import (
 	"github.com/jensgreen/dux/files"
-	"github.com/jensgreen/dux/r1"
-	"github.com/jensgreen/dux/r2"
+	"github.com/jensgreen/dux/geo/r1"
+	"github.com/jensgreen/dux/geo/r2"
 )
 
 // Tiles that are too small to be meaningfully represented individually are put in a
