+++ treemap/treemap.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/jensgreen/dux/files"
 	"github.com/jensgreen/dux/geo"
-	"github.com/jensgreen/dux/r2"
+	"github.com/jensgreen/dux/geo/r2"
 	"github.com/jensgreen/dux/treemap/tiling"
 )
